ajout image, like et dislike bouton, access pour role guest

+++ data/handlers.go
@@ -2,9 +2,12 @@ package data
 
 import (
 	"database/sql"
+	"fmt"
 	"html/template"
+	"io"
 	"log"
 	"net/http"
+	"os"
 	"path/filepath"
 	"strings"
 	"time"
@@ -120,19 +123,13 @@ func Post(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	data := struct {
-		Categories []string
-	}{
-		Categories: categories,
-	}
-
 	if r.Method == "POST" {
 		category := r.FormValue("categories")
 		message := r.FormValue("message")
 
 		session, err := validateSession(r)
 		if err != nil {
-			http.Error(w, "You must be logged in to post", http.StatusUnauthorized)
+			http.Error(w, "Session invalide ou expirée", http.StatusUnauthorized)
 			return
 		}
 
@@ -144,23 +141,75 @@ func Post(w http.ResponseWriter, r *http.Request) {
 		}
 		defer db.Close()
 
-		var user_ID int
-		err = db.QueryRow("SELECT id FROM users WHERE username = ?", session.Username).Scan(&user_ID)
+		var userID int
+		err = db.QueryRow("SELECT id FROM users WHERE username = ?", session.Username).Scan(&userID)
 		if err != nil {
 			log.Println(err)
 			http.Error(w, "Error retrieving user ID", http.StatusInternalServerError)
 			return
 		}
 
-		_, err = db.Exec("INSERT INTO posts (user_id, category_id, content) VALUES (?, ?, ?)", user_ID, category, message)
+		err = os.MkdirAll("uploads", os.ModePerm)
 		if err != nil {
 			log.Println(err)
+			http.Error(w, "Error creating uploads directory", http.StatusInternalServerError)
+			return
+		}
+
+		file, header, err := r.FormFile("image")
+		if err != nil && err != http.ErrMissingFile {
+			log.Println(err)
+			http.Error(w, "Error retrieving file", http.StatusInternalServerError)
+			return
+		}
+
+		var filepath string
+		if err != http.ErrMissingFile {
+			defer file.Close()
+
+			if header.Size > 20*1024*1024 {
+				http.Error(w, "File too large", http.StatusBadRequest)
+				return
+			}
+
+			fileType := header.Header.Get("Content-Type")
+			if fileType != "image/jpeg" && fileType != "image/png" && fileType != "image/gif" {
+				http.Error(w, "Unsupported file type", http.StatusBadRequest)
+				return
+			}
+
+			filename := fmt.Sprintf("%d-%s", userID, header.Filename)
+			filepath = fmt.Sprintf("uploads/%s", filename)
+			outFile, err := os.Create(filepath)
+			if err != nil {
+				log.Println(err)
+				http.Error(w, "Error saving file", http.StatusInternalServerError)
+				return
+			}
+			defer outFile.Close()
+			_, err = io.Copy(outFile, file)
+			if err != nil {
+				log.Println(err)
+				http.Error(w, "Error writing file", http.StatusInternalServerError)
+				return
+			}
+		}
+
+		_, err = db.Exec("INSERT INTO posts (user_id, category_id, content, image_path) VALUES (?, ?, ?, ?)", userID, category, message, filepath)
+		if err != nil {
+			log.Println("Error inserting post:", err)
 			http.Error(w, "Error inserting post", http.StatusInternalServerError)
 			return
 		}
 
 		http.Redirect(w, r, "/home", http.StatusSeeOther)
 		return
+	}
+
+	data := struct {
+		Categories []string
+	}{
+		Categories: categories,
 	}
 
 	tmpl := template.Must(template.ParseFiles("src/templates/post.html"))
@@ -183,20 +232,19 @@ func Categories(w http.ResponseWriter, r *http.Request) {
 	defer db.Close()
 
 	session, err := validateSession(r)
-	if err != nil {
-		http.Error(w, "Session invalide ou expirée", http.StatusUnauthorized)
-		return
-	}
-
 	var currentUserRole string
-	err = db.QueryRow("SELECT Role FROM users WHERE username = ?", session.Username).Scan(&currentUserRole)
 	if err != nil {
-		log.Println(err)
-		http.Error(w, "Error retrieving user role", http.StatusInternalServerError)
-		return
+		currentUserRole = "Guest"
+	} else {
+		err = db.QueryRow("SELECT Role FROM users WHERE username = ?", session.Username).Scan(&currentUserRole)
+		if err != nil {
+			log.Println(err)
+			http.Error(w, "Error retrieving user role", http.StatusInternalServerError)
+			return
+		}
 	}
 
-	if r.Method == "POST" {
+	if r.Method == "POST" && currentUserRole == "Administrators" {
 		categoryID := r.FormValue("category-id")
 		delete := r.FormValue("_method")
 
@@ -205,6 +253,12 @@ func Categories(w http.ResponseWriter, r *http.Request) {
 			if err != nil {
 				log.Println(err)
 				http.Error(w, "Error deleting category", http.StatusInternalServerError)
+				return
+			}
+
+			_, err = db.Exec("DELETE FROM posts WHERE category_id = ?", categoryID)
+			if err != nil {
+				http.Error(w, "Error deleting post", http.StatusInternalServerError)
 				return
 			}
 		}
@@ -258,6 +312,16 @@ func About(w http.ResponseWriter, r *http.Request) {
 	tmpl.Execute(w, errorAbout)
 }
 
+type UserData struct {
+	Username        string
+	Email           string
+	Posts           []string
+	Commentaires    []string
+	Date            []string
+	Auth            string
+	IsAuthenticated bool
+}
+
 func Users(w http.ResponseWriter, r *http.Request) {
 	db, err := sql.Open("sqlite3", "./database.db")
 	if err != nil {
@@ -269,7 +333,14 @@ func Users(w http.ResponseWriter, r *http.Request) {
 
 	session, err := validateSession(r)
 	if err != nil {
-		http.Error(w, "Session invalide ou expirée", http.StatusUnauthorized)
+		data := UserData{
+			Auth:            "You have to be connected to see your profile",
+			IsAuthenticated: false,
+		}
+		tmpl := template.Must(template.ParseFiles("src/templates/user.html"))
+		if err := tmpl.Execute(w, data); err != nil {
+			http.Error(w, "Erreur lors du rendu de la page", http.StatusInternalServerError)
+		}
 		return
 	}
 
@@ -380,18 +451,14 @@ func Users(w http.ResponseWriter, r *http.Request) {
 	}
 
 	tmpl := template.Must(template.ParseFiles("src/templates/user.html"))
-	data := struct {
-		Username     string
-		Email        string
-		Posts        []string
-		Commentaires []string
-		Date         []string
-	}{
-		Username:     username,
-		Email:        email,
-		Posts:        posts,
-		Commentaires: coms,
-		Date:         localDates,
+	data := UserData{
+		Username:        username,
+		Email:           email,
+		Posts:           posts,
+		Commentaires:    coms,
+		Date:            localDates,
+		Auth:            "",
+		IsAuthenticated: true,
 	}
 	if err := tmpl.Execute(w, data); err != nil {
 		http.Error(w, "Erreur lors du rendu de la page", http.StatusInternalServerError)
@@ -405,31 +472,43 @@ func Parameter(w http.ResponseWriter, r *http.Request) {
 
 func Categopost(w http.ResponseWriter, r *http.Request) {
 	categoryName := strings.TrimPrefix(r.URL.Path, "/categories/")
-	posts, postIDs, userIDs, err := getPostsByCategory(categoryName)
+	posts, postIDs, userIDs, imagePaths, dates, err := getPostsByCategory(categoryName)
 	if err != nil {
+		log.Println("Error retrieving posts:", err)
 		http.Error(w, "Error retrieving posts", http.StatusInternalServerError)
 		return
 	}
 
 	session, err := validateSession(r)
-	if err != nil {
-		http.Error(w, "Session invalide ou expirée", http.StatusUnauthorized)
-		return
-	}
-
-	db, err := sql.Open("sqlite3", "./database.db")
-	if err != nil {
-		http.Error(w, "Database connection error", http.StatusInternalServerError)
-		return
-	}
-	defer db.Close()
-
 	var currentUserRole string
-	err = db.QueryRow("SELECT Role FROM users WHERE username = ?", session.Username).Scan(&currentUserRole)
 	if err != nil {
-		log.Println(err)
-		http.Error(w, "Error retrieving user role", http.StatusInternalServerError)
-		return
+		currentUserRole = "Guest"
+	} else {
+		db, err := sql.Open("sqlite3", "./database.db")
+		if err != nil {
+			log.Println("Database connection error:", err)
+			http.Error(w, "Database connection error", http.StatusInternalServerError)
+			return
+		}
+		defer db.Close()
+
+		err = db.QueryRow("SELECT Role FROM users WHERE username = ?", session.Username).Scan(&currentUserRole)
+		if err != nil {
+			log.Println("Error retrieving user role:", err)
+			http.Error(w, "Error retrieving user role", http.StatusInternalServerError)
+			return
+		}
+	}
+
+	localDates := make([]string, len(dates))
+	for i, date := range dates {
+		localTime, err := time.Parse(time.RFC3339, date)
+		if err != nil {
+			log.Println("Error parsing date:", err)
+			localDates[i] = date
+		} else {
+			localDates[i] = localTime.Format("02-01-2006 15:04:05")
+		}
 	}
 
 	var usernames []string
@@ -440,8 +519,7 @@ func Categopost(w http.ResponseWriter, r *http.Request) {
 			http.Error(w, "Error retrieving username", http.StatusInternalServerError)
 			return
 		}
-		name := strings.Join(username, " ")
-		usernames = append(usernames, name)
+		usernames = append(usernames, strings.Join(username, " "))
 	}
 
 	var allCommentaires [][]string
@@ -470,8 +548,7 @@ func Categopost(w http.ResponseWriter, r *http.Request) {
 				http.Error(w, "Error retrieving username", http.StatusInternalServerError)
 				return
 			}
-			name := strings.Join(username, " ")
-			commentUsernames = append(commentUsernames, name)
+			commentUsernames = append(commentUsernames, strings.Join(username, " "))
 		}
 
 		allCommentaires = append(allCommentaires, comments)
@@ -479,33 +556,49 @@ func Categopost(w http.ResponseWriter, r *http.Request) {
 	}
 
 	if r.Method == "POST" {
-		postID := r.FormValue("post-id")
-		content := r.FormValue("reply-message")
-		delete := r.FormValue("_method")
+		if currentUserRole != "Guest" {
+			db, err := sql.Open("sqlite3", "./database.db")
+			if err != nil {
+				log.Println("Database connection error:", err)
+				http.Error(w, "Database connection error", http.StatusInternalServerError)
+				return
+			}
+			defer db.Close()
 
-		var userID int
-		err = db.QueryRow("SELECT id FROM users WHERE username = ?", session.Username).Scan(&userID)
-		if err != nil {
-			http.Error(w, "Error retrieving user ID", http.StatusInternalServerError)
+			postID := r.FormValue("post-id")
+			content := r.FormValue("reply-message")
+			delete := r.FormValue("_method")
+
+			var userID int
+			err = db.QueryRow("SELECT id FROM users WHERE username = ?", session.Username).Scan(&userID)
+			if err != nil {
+				log.Println("Error retrieving user ID:", err)
+				http.Error(w, "Error retrieving user ID", http.StatusInternalServerError)
+				return
+			}
+
+			if delete == "DELETE" && currentUserRole == "Administrators" {
+				_, err = db.Exec("DELETE FROM posts WHERE ID = ?", postID)
+				if err != nil {
+					log.Println("Error deleting post:", err)
+					http.Error(w, "Error deleting post", http.StatusInternalServerError)
+					return
+				}
+			} else if delete != "DELETE" {
+				_, err = db.Exec("INSERT INTO commentaries (postID, user_ID, content) VALUES (?, ?, ?)", postID, userID, content)
+				if err != nil {
+					log.Println("Error inserting commentary:", err)
+					http.Error(w, "Error inserting commentary", http.StatusInternalServerError)
+					return
+				}
+			}
+
+			http.Redirect(w, r, r.URL.Path, http.StatusSeeOther)
+			return
+		} else {
+			http.Error(w, "Unauthorized action", http.StatusUnauthorized)
 			return
 		}
-
-		if delete == "DELETE" {
-			_, err = db.Exec("DELETE FROM posts WHERE ID = ?", postID)
-			if err != nil {
-				http.Error(w, "Error deleting post", http.StatusInternalServerError)
-				return
-			}
-		} else {
-			_, err = db.Exec("INSERT INTO commentaries (postID, user_ID, content) VALUES (?, ?, ?)", postID, userID, content)
-			if err != nil {
-				http.Error(w, "Error inserting commentary", http.StatusInternalServerError)
-				return
-			}
-		}
-
-		http.Redirect(w, r, "/categories", http.StatusSeeOther)
-		return
 	}
 
 	tmplPath := filepath.Join("src/templates/categopost.html")
@@ -523,7 +616,9 @@ func Categopost(w http.ResponseWriter, r *http.Request) {
 		Usernames          []string
 		Commentaires       [][]string
 		AuteurCommentaires [][]string
+		ImagePaths         []string
 		CurrentUserRole    string
+		localDates         []string
 	}{
 		Category:           categoryName,
 		Posts:              posts,
@@ -531,7 +626,9 @@ func Categopost(w http.ResponseWriter, r *http.Request) {
 		Usernames:          usernames,
 		Commentaires:       allCommentaires,
 		AuteurCommentaires: allAuteurCommentaires,
+		ImagePaths:         imagePaths,
 		CurrentUserRole:    currentUserRole,
+		localDates:         localDates,
 	}
 
 	err = tmpl.Execute(w, data)
@@ -567,6 +664,27 @@ func OtherProfile(w http.ResponseWriter, r *http.Request) {
 		} else {
 			localDates[i] = localTime.Format("02-01-2006 15:04:05")
 		}
+	}
+
+	session, err := validateSession(r)
+	if err != nil {
+		http.Error(w, "Session invalide ou expirée", http.StatusUnauthorized)
+		return
+	}
+
+	db, err := sql.Open("sqlite3", "./database.db")
+	if err != nil {
+		http.Error(w, "Database connection error", http.StatusInternalServerError)
+		return
+	}
+	defer db.Close()
+
+	var currentUserRole string
+	err = db.QueryRow("SELECT Role FROM users WHERE username = ?", session.Username).Scan(&currentUserRole)
+	if err != nil {
+		log.Println(err)
+		http.Error(w, "Error retrieving user role", http.StatusInternalServerError)
+		return
 	}
 
 	if r.Method == "POST" {
@@ -605,15 +723,17 @@ func OtherProfile(w http.ResponseWriter, r *http.Request) {
 
 	tmpl := template.Must(template.ParseFiles("src/templates/otherprofile.html"))
 	data := struct {
-		Dates    []string
-		Post     []string
-		Username string
-		Status   []string
+		Dates           []string
+		Post            []string
+		Username        string
+		Status          []string
+		CurrentUserRole string
 	}{
-		Dates:    localDates,
-		Post:     content,
-		Username: OtUsername,
-		Status:   roles,
+		Dates:           localDates,
+		Post:            content,
+		Username:        OtUsername,
+		Status:          roles,
+		CurrentUserRole: currentUserRole,
 	}
 	if err := tmpl.Execute(w, data); err != nil {
 		http.Error(w, "Erreur lors du rendu de la page", http.StatusInternalServerError)
